backend/requirements.txt
# pip install -r requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.16
sqlalchemy==2.0.30
psycopg[binary]==3.1.19
alembic==1.13.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
email-validator==2.1.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pandas==2.2.2
pgvector==0.4.2
openai==2.24.0